Refactor chars and spaces into its own module
